+++ services/iaas/model_create_affinity_group_payload.go
@@ -26,7 +26,7 @@ type CreateAffinityGroupPayload struct {
 	// The name for a General Object. Matches Names and also UUIDs.
 	// REQUIRED
 	Name *string `json:"name"`
-	// The affinity group policy.
+	// The affinity group policy. `hard-affinity`: All servers in this group will be hosted on the same compute node. `soft-affinity`: All servers in this group will be hosted on as few compute nodes as possible. `hard-anti-affinity`: All servers in this group will be hosted on different compute nodes. `soft-anti-affinity`: All servers in this group will be hosted on as many compute nodes as possible. Possible values: `hard-anti-affinity`, `hard-affinity`, `soft-anti-affinity`, `soft-affinity`.
 	// REQUIRED
 	Policy *string `json:"policy"`
 }