+++ services/iaas/model_update_nic_payload.go
@@ -21,7 +21,7 @@ var _ MappedNullable = &UpdateNicPayload{}
 type UpdateNicPayload struct {
 	// A list of IPs or CIDR notations.
 	AllowedAddresses *[]AllowedAddressesInner `json:"allowedAddresses,omitempty"`
-	// Object that represents the labels of an object.
+	// Object that represents the labels of an object. Regex for keys: `^[a-z]((-|_|[a-z0-9])){0,62}$`. Regex for values: `^(-|_|[a-z0-9]){0,63}$`.
 	Labels *map[string]interface{} `json:"labels,omitempty"`
 	// The name for a General Object. Matches Names and also UUIDs.
 	Name *string `json:"name,omitempty"`